+++ src/main/java/hello/backend/error/exception/user/BadRequestException.java
@@ -1,4 +1,4 @@
-package hello.backend.error.exception;
+package hello.backend.error.exception.user;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
